Do not blanket delete events

+++ exporter.py
@@ -4,6 +4,7 @@ import json
 import datetime
 import csv
 import io
+import re
 from icalendar import Calendar, Event, vText
 import caldav
 
@@ -12,6 +13,7 @@ QUERY_PARAMS = {
     "worklogTimeExport": "seconds"
 }
 
+WORKLOG_ID_REGEX = re.compile(r'\[JIRA:(\S+)\]')
 START_TIME_FORMAT = '%d. %b %Y %H:%M'
 
 START_TIME_KEY = "Start Time"
@@ -19,13 +21,17 @@ TIME_SPENT_KEY = "Time Spent (s)"
 WORKLOG_DESCRIPTION_KEY = "Worklog Description"
 ISSUE_NUMBER_KEY = "Issue Key"
 ISSUE_TITLE_KEY = "Issue Summary"
+WORKLOG_ID_KEY = "Worklog ID"
+
+BUCKET_ISSUE_KEYS = ["ALDE-2", "ALDE-3"]
 
 KEYS = [
     START_TIME_KEY,
     TIME_SPENT_KEY,
     WORKLOG_DESCRIPTION_KEY,
     ISSUE_NUMBER_KEY,
-    ISSUE_TITLE_KEY
+    ISSUE_TITLE_KEY,
+    WORKLOG_ID_KEY,
 ]
 
 DAYS_TO_PROCESS = 3
@@ -66,25 +72,41 @@ def parse_csv(csv_text):
     return events
 
 
-def to_ical(events):
-    cal = Calendar()
-
-    for event in events:
-        start_time = datetime.datetime.strptime(event[START_TIME_KEY], START_TIME_FORMAT)
-        seconds = int(event[TIME_SPENT_KEY])
-        end_time = start_time + datetime.timedelta(seconds=seconds)
-
-        cal_event = Event()
-        cal_event.add('summary', f'{event[ISSUE_NUMBER_KEY]}: {event[WORKLOG_DESCRIPTION_KEY]}')
-        cal_event.add('dtstart', start_time), 
-        cal_event.add('dtend', end_time), 
-        cal_event['description'] = vText(f"{event[ISSUE_NUMBER_KEY]}: {event[ISSUE_TITLE_KEY]}")
-        cal.add_component(cal_event)
-    
-    return cal.to_ical().decode("utf-8")
+def create_event_properties(jira_event):
+    short_desc = jira_event[WORKLOG_DESCRIPTION_KEY].splitlines()[0]
+    start_time = datetime.datetime.strptime(jira_event[START_TIME_KEY], START_TIME_FORMAT)
+    seconds = int(jira_event[TIME_SPENT_KEY])
+    end_time = start_time + datetime.timedelta(seconds=seconds)
+    summary = f'{jira_event[ISSUE_NUMBER_KEY]}: {short_desc}' \
+        if jira_event[ISSUE_NUMBER_KEY] not in BUCKET_ISSUE_KEYS \
+        else f'{short_desc}'
+    return {
+        "dtstart": start_time,
+        "dtend": end_time,
+        "summary": summary,
+        "description": f"{jira_event[ISSUE_NUMBER_KEY]}: {jira_event[ISSUE_TITLE_KEY]}\n\n{jira_event[WORKLOG_DESCRIPTION_KEY]}\n\n[JIRA:{jira_event[WORKLOG_ID_KEY]}]"
+    }
 
 
-def push_to_caldav(events, url, user, password, name, fromtime, totime):
+def update_event(caldav_event, jira_event):
+    props = create_event_properties(jira_event)
+    for key, val in props.items():
+        getattr(caldav_event.vobject_instance.vevent, key).value = val
+    caldav_event.save()
+
+
+def find_matching_caldav_event(jira_event, caldav_events):
+    for e in caldav_events:
+        alldata = "".join(e.data.split())
+        match = WORKLOG_ID_REGEX.search(alldata)
+        if match:
+            existing_id = match.group(1)
+            if existing_id == jira_event[WORKLOG_ID_KEY]:
+                return e
+    return None
+
+
+def push_to_caldav(events, url, user, password, name, fromtime, totime, wipe):
     with caldav.DAVClient(
         url=url,
         username=user,
@@ -101,22 +123,17 @@ def push_to_caldav(events, url, user, password, name, fromtime, totime):
             expand=False,
         )
 
-        for e in calendar_events:
-            e.delete()
+        if wipe:
+            for e in calendar_events:
+                e.delete()
+            calendar_events = []
+
         for event in events:
-            short_desc = event[WORKLOG_DESCRIPTION_KEY].splitlines()[0]
-            start_time = datetime.datetime.strptime(event[START_TIME_KEY], START_TIME_FORMAT)
-            seconds = int(event[TIME_SPENT_KEY])
-            end_time = start_time + datetime.timedelta(seconds=seconds)
-            summary = f'{event[ISSUE_NUMBER_KEY]}: {short_desc}' \
-                if event[ISSUE_NUMBER_KEY] not in ["ALDE-2", "ALDE-3"] \
-                else f'{short_desc}'
-            calendar.save_event(
-                dtstart=start_time,
-                dtend=end_time,
-                summary=summary,
-                description=f"{event[ISSUE_NUMBER_KEY]}: {event[ISSUE_TITLE_KEY]}\n\n{event[WORKLOG_DESCRIPTION_KEY]}"
-            )
+            c = find_matching_caldav_event(event, calendar_events)
+            if c:
+                update_event(c, event)
+            else:
+                calendar.save_event(**create_event_properties(event))
 
 
 def main(opts):
@@ -128,7 +145,7 @@ def main(opts):
     csv_text = get_worklogs(opts.token, opts.reportfilter, past_timestamp, today_timestamp, opts.jira)
     events = parse_csv(csv_text)
 
-    push_to_caldav(events, opts.url, opts.user, opts.password, opts.calendar, past, today)
+    push_to_caldav(events, opts.url, opts.user, opts.password, opts.calendar, past, today, opts.wipe)
 
 
 if __name__ == "__main__":
@@ -142,6 +159,7 @@ if __name__ == "__main__":
     parser.add_option('-p', '--caldav-pass', dest='password', type='string', help="CalDAV password")
     parser.add_option('-n', '--calendar-name', dest='calendar', type='string', help="CalDAV calendar name")
     parser.add_option('-d', '--days', dest='days', type='int', help="How many days to go back in time", default=DAYS_TO_PROCESS)
+    parser.add_option('-w', '--wipe', dest='wipe', action='store_true', help="If set, remove all existing events within time range from target calendar", default=False)
 
     (opts, args) = parser.parse_args()
     main(opts)
